# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KaiPOS API",
    "version": "1.0.0",
    "description": "KaiPOS backend HTTP API. Generated from Zod schemas via `pnpm --filter @kaipos/backend openapi:generate`."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "pnpm dev (Docker Mongo)"
    },
    {
      "url": "http://localhost:4001",
      "description": "pnpm docker:up (Docker Mongo)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["products"],
        "summary": "List products with optional search/filter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "branchId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": false,
            "name": "includeInactive",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": false,
            "name": "activeNow",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "featuredIn",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "businessId",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          }
                        },
                        "required": ["_id"],
                        "additionalProperties": {
                          "nullable": true
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "limit": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "totalPages": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": ["page", "limit", "total", "totalPages"]
                    }
                  },
                  "required": ["success", "data", "pagination"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["products"],
        "summary": "Create a product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0
                  },
                  "category": {
                    "type": "string",
                    "minLength": 1
                  },
                  "sku": {
                    "type": "string",
                    "minLength": 1
                  },
                  "stock": {
                    "type": "number",
                    "minimum": 0
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "cost": {
                    "type": "number",
                    "minimum": 0
                  },
                  "taxRate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "trackStock": {
                    "type": "boolean",
                    "default": true
                  },
                  "lowStockThreshold": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "stockUnit": {
                    "type": "string",
                    "enum": ["unit", "kg", "L"],
                    "default": "unit"
                  },
                  "availability": {
                    "type": "object",
                    "properties": {
                      "pos": {
                        "type": "boolean"
                      },
                      "online": {
                        "type": "boolean"
                      },
                      "kiosk": {
                        "type": "boolean"
                      }
                    },
                    "default": {
                      "pos": true,
                      "online": false,
                      "kiosk": false
                    },
                    "required": ["pos", "online", "kiosk"]
                  },
                  "serviceSchedules": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["breakfast", "lunch", "dinner"]
                    },
                    "default": []
                  },
                  "allergens": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "gluten",
                        "dairy",
                        "egg",
                        "peanut",
                        "tree-nut",
                        "soy",
                        "fish",
                        "shellfish",
                        "sesame"
                      ]
                    },
                    "default": []
                  },
                  "dietaryTags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["vegetarian", "vegan", "gluten-free", "keto", "halal", "kosher"]
                    },
                    "default": []
                  },
                  "modifierGroups": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "required": {
                          "type": "boolean"
                        },
                        "maxSelectable": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "options": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "minLength": 1
                              },
                              "label": {
                                "type": "string",
                                "minLength": 1
                              },
                              "priceDelta": {
                                "type": "number"
                              },
                              "available": {
                                "type": "object",
                                "properties": {
                                  "daysOfWeek": {
                                    "type": "array",
                                    "items": {
                                      "type": "integer",
                                      "minimum": 0,
                                      "maximum": 6
                                    },
                                    "minItems": 1
                                  },
                                  "from": {
                                    "type": "string",
                                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                                  },
                                  "to": {
                                    "type": "string",
                                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                                  }
                                },
                                "additionalProperties": false
                              }
                            },
                            "required": ["id", "label", "priceDelta"]
                          }
                        }
                      },
                      "required": ["id", "name", "required", "maxSelectable", "options"]
                    },
                    "default": []
                  },
                  "kitchenStationIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "default": []
                  },
                  "variants": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "sku": {
                          "type": "string",
                          "minLength": 1
                        },
                        "priceDelta": {
                          "type": "number"
                        },
                        "imageUrl": {
                          "type": "string",
                          "format": "uri"
                        }
                      },
                      "required": ["id", "name", "sku", "priceDelta"]
                    }
                  },
                  "availabilityWindow": {
                    "type": "object",
                    "properties": {
                      "daysOfWeek": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 6
                        },
                        "minItems": 1
                      },
                      "from": {
                        "type": "string",
                        "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                      },
                      "to": {
                        "type": "string",
                        "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                      }
                    },
                    "required": ["daysOfWeek", "from", "to"]
                  },
                  "sortOrder": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 0
                  },
                  "barcode": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["branchId", "name", "description", "price", "category", "sku", "stock"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        }
                      },
                      "required": ["_id"],
                      "additionalProperties": {
                        "nullable": true
                      }
                    }
                  },
                  "required": ["success", "data"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/reorder": {
      "patch": {
        "tags": ["products"],
        "summary": "Bulk-update sortOrder for products in a branch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "sortOrder": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": ["id", "sortOrder"]
                    },
                    "minItems": 1,
                    "maxItems": 500
                  }
                },
                "required": ["branchId", "items"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reorder applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "matched": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": ["matched"]
                    }
                  },
                  "required": ["success", "data"]
                }
              }
            }
          },
          "400": {
            "description": "One or more product ids not found in this branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["success", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/preferences": {
      "get": {
        "tags": ["products"],
        "summary": "Featured product ids for a branch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "branchId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ids of the products featured in this branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "featuredProductIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": ["featuredProductIds"]
                    }
                  },
                  "required": ["success", "data"]
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["success", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}/feature": {
      "patch": {
        "tags": ["products"],
        "summary": "Toggle a product as featured for a branch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "featured": {
                    "type": "boolean"
                  }
                },
                "required": ["branchId", "featured"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated product preference",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        }
                      },
                      "required": ["_id"],
                      "additionalProperties": {
                        "nullable": true
                      }
                    }
                  },
                  "required": ["success", "data"]
                }
              }
            }
          },
          "404": {
            "description": "Product not found in this branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["success", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/upload-url": {
      "post": {
        "tags": ["products"],
        "summary": "Pre-signed S3 URL for product image upload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "contentType": {
                    "type": "string",
                    "enum": ["image/jpeg", "image/png", "image/webp"]
                  },
                  "fileSize": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 2097152
                  }
                },
                "required": ["branchId", "contentType", "fileSize"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pre-signed PUT URL + public URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "uploadUrl": {
                          "type": "string",
                          "format": "uri"
                        },
                        "publicUrl": {
                          "type": "string",
                          "format": "uri"
                        }
                      },
                      "required": ["uploadUrl", "publicUrl"]
                    }
                  },
                  "required": ["success", "data"]
                }
              }
            }
          },
          "503": {
            "description": "Asset bucket not configured for this environment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["success", "error"]
                }
              }
            }
          }
        }
      }
    }
  }
}