# apps/backend/src/routes/products.ts
import { Hono } from 'hono';
import type { AppEnv } from '../types.js';
import { validate } from '../middleware/validation.js';
import { requireAuth } from '../middleware/auth.js';
import { requirePermission } from '../middleware/authorize.js';
import { requireBranchAccess } from '../middleware/branch-access.js';
import {
  createProductSchema,
  featureProductSchema,
  listProductPreferencesQuerySchema,
  listProductsQuerySchema,
  productIdParamSchema,
  reorderProductsSchema,
  updateProductSchema,
  uploadUrlSchema,
  type CreateProductInput,
  type FeatureProductInput,
  type ListProductsQuery,
  type ReorderProductsInput,
  type UpdateProductInput,
  type UploadUrlInput,
} from '../schemas/products.js';
import * as productsService from '../services/products.js';
import * as productPreferencesService from '../services/product-preferences.js';
import { toPaginatedResponse } from '../lib/paginate.js';

const products = new Hono<AppEnv>();

products.post(
  '/api/products/upload-url',
  requireAuth(),
  requirePermission('products:write'),
  validate({ body: uploadUrlSchema }),
  async (c) => {
    const user = c.get('user')!;
    const body = (await c.req.json()) as UploadUrlInput;
    const result = await productsService.generateUploadUrl(user, body);
    return c.json({ success: true, data: result }, 201);
  },
);

products.get(
  '/api/products',
  requireAuth(),
  requirePermission('products:read'),
  validate({ query: listProductsQuerySchema }),
  requireBranchAccess('branchId', 'query'),
  async (c) => {
    const user = c.get('user')!;
    const parsed = listProductsQuerySchema.parse(c.req.query()) as ListProductsQuery;
    const result = await productsService.listProducts(user, parsed);
    return c.json(toPaginatedResponse(result));
  },
);

// Specific paths must be registered before `/:id` so they don't get matched
// as a product id.
products.patch(
  '/api/products/reorder',
  requireAuth(),
  requirePermission('products:write'),
  validate({ body: reorderProductsSchema }),
  requireBranchAccess('branchId', 'body'),
  async (c) => {
    const user = c.get('user')!;
    const body = reorderProductsSchema.parse(await c.req.json()) as ReorderProductsInput;
    const result = await productsService.reorderProducts(user, body);
    return c.json({ success: true, data: result });
  },
);

// Featured product ids for a branch. Registered before `/:id` so it isn't
// matched as a product whose id is "preferences".
products.get(
  '/api/products/preferences',
  requireAuth(),
  requirePermission('products:read'),
  validate({ query: listProductPreferencesQuerySchema }),
  requireBranchAccess('branchId', 'query'),
  async (c) => {
    const user = c.get('user')!;
    const { branchId } = listProductPreferencesQuerySchema.parse(c.req.query());
    const featuredProductIds = await productPreferencesService.listFeaturedProductIds(
      user,
      branchId,
    );
    return c.json({ success: true, data: { featuredProductIds } });
  },
);

products.get(
  '/api/products/:id',
  requireAuth(),
  requirePermission('products:read'),
  validate({ params: productIdParamSchema }),
  async (c) => {
    const user = c.get('user')!;
    const id = c.req.param('id');
    const result = await productsService.getProductById(user, id);
    return c.json({ success: true, data: result });
  },
);

products.post(
  '/api/products',
  requireAuth(),
  requirePermission('products:write'),
  validate({ body: createProductSchema }),
  async (c) => {
    const user = c.get('user')!;
    const body = createProductSchema.parse(await c.req.json()) as CreateProductInput;
    const result = await productsService.createProduct(user, body);
    return c.json({ success: true, data: result }, 201);
  },
);

products.patch(
  '/api/products/:id/feature',
  requireAuth(),
  requirePermission('products:write'),
  validate({ params: productIdParamSchema, body: featureProductSchema }),
  requireBranchAccess('branchId', 'body'),
  async (c) => {
    const user = c.get('user')!;
    const id = c.req.param('id');
    const body = featureProductSchema.parse(await c.req.json()) as FeatureProductInput;
    const result = await productPreferencesService.setFeatured(user, id, body);
    return c.json({ success: true, data: result });
  },
);

products.patch(
  '/api/products/:id',
  requireAuth(),
  requirePermission('products:write'),
  validate({ params: productIdParamSchema, body: updateProductSchema }),
  async (c) => {
    const user = c.get('user')!;
    const id = c.req.param('id');
    const body = updateProductSchema.parse(await c.req.json()) as UpdateProductInput;
    const result = await productsService.updateProduct(user, id, body, {
      route: c.req.path,
      method: c.req.method,
    });
    return c.json({ success: true, data: result });
  },
);

products.delete(
  '/api/products/:id',
  requireAuth(),
  requirePermission('products:delete'),
  validate({ params: productIdParamSchema }),
  async (c) => {
    const user = c.get('user')!;
    const id = c.req.param('id');
    await productsService.deleteProduct(user, id, {
      route: c.req.path,
      method: c.req.method,
    });
    return c.body(null, 204);
  },
);

export default products;